requirements.txt
arrow==0.8.0
beautifulsoup4==4.5.1
blessed==1.14.1
django-crontab==0.7.1
django-environ==0.4.1
django-picklefield==0.3.2
future==0.16.0
py==1.4.31
pytest==3.0.5
pytest-django==3.1.2
python-dateutil==2.6.0
redis==2.10.5
selenium==3.0.2
six==1.10.0
wcwidth==0.1.7
